requirements.txt
canvasapi==3.3.0
ticktick-py==2.0.8
python-dotenv==1.0.0
PyYAML==6.0.1
beautifulsoup4==4.12.3
playwright==1.49.0
